@@ -1,6 +1,5 @@
-const httpClient = require('./httpClient');
-
-var buildErCqlFilter = require('./buildErCqlFilter');
+import httpClient from './httpClient.js';
+import buildErCqlFilter from './buildErCqlFilter.js';
 
 /**
  * @classdesc
@@ -105,4 +104,4 @@ ClientEr.prototype.getFeatures = function (typeName, params) {
     });
 };
 
-module.exports = ClientEr;
+export {ClientEr};
