@@ -1,4 +1,7 @@
-{
+import globals from 'globals';
+
+export default [
+    {
     "rules": {
         "indent": [
             2,
@@ -20,9 +23,12 @@
             1
         ]
     },
-    "env": {
-        "es6": true,
-        "node": true
-    },
-    "extends": "eslint:recommended"
-}
+    languageOptions: {
+        sourceType: "module",
+        globals: {
+          ...globals.node,
+          ...globals.browser
+        }
+    }
+}   
+]
